AoE Enchant Power Transfer Commit
Added the feature for AoE Enchantments to transfer their damage to another already existing damage effect rather than create a new one.

This functionality exists for Quake, Aberration, Eruption, Soulfire, and Starfall.

For Aberration, this will only activate when there are already 4 slime entities within 16 blocks.

For Starfall, this will only activate when there are already 2 meteors within 12 blocks.

For all the other enchantments, this will happen if there is already one instance of the effect within 12 blocks (16 blocks for Soulfire).

When power is transferred, the already existing AoE effect will absorb the damage of the new effect, instead of creating the new effect. This will only happen if the corresponding effect is not dealing any damage. Particles are shown when this occurs as well as a unique sound.

Balance adjustments:
Quake's default size has also been reduced from a 6 block radius to a 4 block radius (12 to 8).