Fix Trident behavior with Infinity, Scavenger, Auto Charge, and Repeating
It is a known and currently will not be fixed bug that these enchantments + trident do not play nice with durability. The tridents are considered consumables meaning that they will not always track durability correctly. A patch or future enchantment for throw rate based tridents will be made in the future to address this.